transferred menu functions to a separate module `menu.py`
